requirements.txt
datasets=1.16.1=pypi_0
en-core-web-sm=3.2.0=pypi_0
huggingface-hub=0.2.1=pypi_0
jsonlines=3.0.0=pypi_0
nltk=3.6.5=pypi_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
pandas=1.3.5=pypi_0
rouge-score=0.0.4=pypi_0
sentence-transformers=2.1.0=pypi_0
sentencepiece=0.1.96=pypi_0
six=1.16.0=pyhd3eb1b0_0
sklearn=0.0=pypi_0
spacy=3.2.1=pypi_0
spacy-legacy=3.0.8=pypi_0
spacy-loggers=1.0.1=pypi_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow-estimator=2.7.0=pypi_0
tensorflow-gpu=2.7.0=pypi_0
tensorflow-io-gcs-filesystem=0.23.1=pypi_0
tokenizers=0.10.3=pypi_0
torchaudio=0.10.1=py38_cu113
torchvision=0.11.2=py38_cu113
tornado=6.1=py38h27cfd23_0
tqdm=4.62.3=pypi_0
transformers=4.14.1=pypi_0